Implement interrupt handling
The esp-idf wifi driver expects to be able to set an ISR at runtime, not
compile-time. esp-idf & freertos allows this, but the rust xtensa
runtime support expects ISRs to be set up at compile-time.  To get
around this, we implement an empty ISR at compile-time, and call the
wifi driver's provided ISR after the driver tells us about it.

This isn't particularly ideal; this might mean it might be a good idea
to add first-class support for runtime ISR changes to the
xtensa-lx/esp32 crates.

Additionally, this implements _queue_send_from_isr() by just forwarding
to the non-ISR impl.  Which may not be safe.