Implement intr_matrix_set()
This is a bit weird: if I just forward the impl to the built-in ROM
routine, we get some sort of livelock where the two cores keep sending
the same data back and forth in the same queue, but nothing happens (and
the wifi AP scan never completes).

The minimal (and probably not-quite-correct) implementation I have here
actually seems to wire up the interrupts correctly.  But of course we
still find zero APs, so...